type: fix
ProfileView의 로컬 db 연동 부분 값 표시 전체 완료

+++ app/src/main/java/com/example/gongik/model/data/myinformation/MyWorkInformation.kt
@@ -10,7 +10,7 @@ data class MyWorkInformation (
     // 복무지
     @ColumnInfo(name = "work_place") val workPlace: String,
     // 소집일
-    @ColumnInfo(name = "start_work_day") val startWorkDay: Int,
+    @ColumnInfo(name = "start_work_day") val startWorkDay: Long,
     // 소집 해제일
-    @ColumnInfo(name = "finish_work_day") val finishWorkDay: Int,
+    @ColumnInfo(name = "finish_work_day") val finishWorkDay: Long,
 )
